refactor(#4174): move flatten-at-materialization helpers to string-materialize.ts
LOC-budget: implementation lives in the new subsystem module; the god-files
keep thin call sites, granted via loc-budget-allow in the issue frontmatter.
